Betting class is separated
