+++ libraries/chain/data_service_factory.cpp
@@ -34,7 +34,6 @@
 #include <scorum/chain/services/blocks_story.hpp>
 #include <scorum/chain/services/advertising_property.hpp>
 #include <scorum/chain/services/betting_property.hpp>
-#include <scorum/chain/services/betting_service.hpp>
 #include <scorum/chain/services/bet.hpp>
 #include <scorum/chain/services/matched_bet.hpp>
 #include <scorum/chain/services/pending_bet.hpp>
@@ -80,7 +79,6 @@ DATA_SERVICE_FACTORY_IMPL(
         (blocks_story)
         (advertising_property)
         (betting_property)
-        (betting)
         (bet)
         (matched_bet)
         (pending_bet)
